Obstacles and Searches Modified
A realistic obstacle grid was set up for a building layout. The
obstacles will now only consist of this realistic setup and a no
obstacle simulation. A hybrid Optimal-Lookahead/Drosophila search was
created in order to take advantages of the lookahead searches local
optimization capabilities, but allowing the Drosophila search to
occasionally ensure exploration/avoid getting stuck on local maxima.