Lookahead Rewritten
The optimal lookahead algorithm has been rewritten after a general
feeling of paranoia struck me concerning the original algorithm. The
search behavior of the original algorithm seemed "off" to me after
watching it for the last week. The new version definitely behaves in the
expected manner, whether or not the old version works the same way has
not yet been tested.